Script Update
Updated to include functionality. Now it can also handle substitute numbers based CLI combined with CSID.  This is useful if multiple senders are sending from the same CLI, which is common for cloud faxing services. We can use both CLI and CSID to determine which number to actually fax back.